refactor: rename festival-event to festival-day
