+++ src/components/afishe-festival-day/index.ts
@@ -1,0 +1,1 @@
+export * from './afishe-festival-day';
