--- src/components/afishe-festival-event/index.ts
@@ -1,1 +1,0 @@
-export * from './afishe-festival-event';
