--- src/components/afishe-festival-events/afishe-festival-events.tsx
@@ -1,15 +1,0 @@
-import {FC} from 'react';
-
-import {FestivalEvent} from '../afishe-festival-event';
-
-import styles from './afishe-festival-events.module.css';
-
-export const FestivalEvents: FC = () => {
-  return (
-    <section className={styles.section}>
-      <FestivalEvent/>
-      <FestivalEvent/>
-      <FestivalEvent/>
-    </section>
-  );
-};
