--- src/components/afishe-festival-events/index.ts
@@ -1,1 +1,0 @@
-export * from './afishe-festival-events';
